abstract indexable behavior to concern + make grade indexable
